+++ Sources/TidalAPIInterchangeKit/API/Models/TidalArtistTracks.swift
@@ -8,18 +8,18 @@
 
 import Foundation
 
-// For the schema this type was based on, see: `https://tidal-music.github.io/tidal-api-reference/tidal-api-oas.json#/components/schemas/Artists_Multi_Relationship_Data_Document`
+// For the schema this type was based on, see: `https://tidal-music.github.io/tidal-api-reference/tidal-api-oas.json#/components/schemas/Artists_Single_Resource_Data_Document`
 
 public struct TidalArtistTracks: Codable, Equatable, Hashable, Sendable {
 
-    public let data: [TidalResource]?
+    public let data: TidalArtist
 
     public let included: [TidalTrack]?
 
     // Not including the links property for now. They are either already known or can be generated from known data.
     // public let links: ...
 
-    public init(data: [TidalResource]? = nil,
+    public init(data: TidalArtist,
                 included: [TidalTrack]? = nil) {
         self.data = data
         self.included = included